A lot of changes.
The detector geometry is now included and called in the EventClass.
The log file is read to extract the nthrown as well as the BField value and the geometry file number.

The StatusHistograms are now turned off by default because they are slowing down Clark due to the wrapping of the histograms.